Separate code for item route to not repeat
