+++ client/app/routes/collection/group/item.js
@@ -1,25 +1,8 @@
 import Ember from 'ember';
+import {itemRoute} from 'collections/utils/itemRoute';
 
-const hasFileList = ['preprint', 'file', 'poster'];
-const hasWikiList = ['project', 'registration'];
 export default Ember.Route.extend({
     model (params){
-      let self = this;
-      return this.store.findRecord('item', params.group_item_id).then(function(item){
-        item.set('hasFile', hasFileList.includes(item.get('type')));
-        item.set('hasWiki', hasWikiList.includes(item.get('type')));
-        self.get('store').findRecord('node', item.get('source_id')).then(function(node){
-            item.set('node', node);
-            node.get('preprints').then(result => {
-                if(result.objectAt(0)){
-                    result.objectAt(0).get('primaryFile').then(pf => {
-                        item.set('downloadUrl', pf.get('links').download);
-                        item.set('hasFile', true);
-                    });
-                }
-             });
-        });
-        return item;
-      });
+        return itemRoute.call(this, params, 'group_item_id');
     }
 });
